chore: remoção do useSignup
